[feature]: инициализация сцены из Settings + правки конфигов и тестов
### #Doc/TODO/todo.md
- Добавлена критичная заметка: одноразовые тесты должны работать с локальными копиями менеджеров.

### #Src/Managers/manager_scene.cpp
- Изменён тип параметров Init на tuple<const std::string, int, int> для совместимости с настройками.

### #Src/Settings/viewport_settings.jsonc
- Обновлён title по умолчанию на "MainViewport" для единообразия с кодом.

### #Src/game_loop.h
- Загрузка VIEWPORT_SETTINGS перед созданием менеджеров.
- Проброс в Scene::Init значений title/width/height из Settings через GetValue.
- Приведены размеры к int в соответствии с сигнатурой Init.

### #Tests/test_manager_scene.h
- Тесты создают и закрывают вьюпорт "TestMainViewport" для избегания коллизий имён.

### #Tests/test_manager_settings.h
- Явный вызов FreeResources() после одноразового прогона тестов.

### #Дополнительно
- Подготовка к изоляции одноразовых тестов, чтобы не изменять глобальное состояние.

+++ Src/Managers/manager_scene.cpp
@@ -14,8 +14,7 @@ void Scene::Init(std::vector<std::any> args) {
   }
   std::tuple<std::string, uint16_t, uint16_t> L_viewport_params;
   try {
-    L_viewport_params = std::any_cast<std::tuple<const std::string, uint16_t, uint16_t>>(
-        args[0]);
+    L_viewport_params = std::any_cast<std::tuple<const std::string, int, int>>(args[0]);
   } catch (const std::bad_any_cast&) {
     LOG::Fatal(__FILE__, __LINE__) << "Scene::Init - invalid viewport parameters";
     return;
